refactor(#131): remove Playwright-based editor operation helpers
Remove all Playwright-based browser automation helpers that have been
replaced by API-based implementations. Content operations on note.com
are now fully API-driven.

Changes:
- Move insert_image_via_api from browser/insert_image.py to api/images.py
- Update server.py import path
- Delete 8 browser helper source files (insert_image, insert_link,
  insert_embed, embed_helpers, image_helpers, toc_helpers,
  text_align_helpers, get_article)
- Delete 9 browser-related test files
- Remove ProseMirrorStabilizer from e2e helpers
- Update test_insert_image.py to use correct mock paths
- Simplify test_insert_body_image.py to API-only tests
- Update CLAUDE.md documentation

Playwright remains in use only for login and preview display.
